Split stores
