+++ web/components/core/Display/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled, { css } from 'styled-components'
 
-import useStore from '@/store'
+import useStore from '@/store/gradient'
 
 interface WrapperProps {
   visible: boolean,
